Corrects Module Reorg
