+++ src/CofreeBot.hs
@@ -1,21 +1,11 @@
 module CofreeBot
-  ( module Bot
-  , module Calculator
+  ( module Behaviors
+  , module Bot
   , module Context
-  , module CoinFlip
-  , module GHCI
-  , module Hello
-  , module Matrix
-  , module Simple
   , module Utils
   ) where
 
 import           CofreeBot.Bot                 as Bot
-import           CofreeBot.Bot.Calculator      as Calculator
-import           CofreeBot.Bot.CoinFlip        as CoinFlip
+import           CofreeBot.Bot.Behaviors       as Behaviors
 import           CofreeBot.Bot.Context         as Context
-import           CofreeBot.Bot.GHCI            as GHCI
-import           CofreeBot.Bot.Hello           as Hello
-import           CofreeBot.Bot.Matrix          as Matrix
-import           CofreeBot.Bot.Simple          as Simple
 import           CofreeBot.Utils               as Utils
