+++ src/CofreeBot/Bot/Behaviors.hs
@@ -1,0 +1,13 @@
+module CofreeBot.Bot.Behaviors
+  ( module Calculator
+  , module CoinFlip
+  , module GHCI
+  , module Hello
+  ) where
+
+import           CofreeBot.Bot.Behaviors.Calculator
+                                               as Calculator
+import           CofreeBot.Bot.Behaviors.CoinFlip
+                                               as CoinFlip
+import           CofreeBot.Bot.Behaviors.GHCI  as GHCI
+import           CofreeBot.Bot.Behaviors.Hello as Hello
